Add class and alter chart code
Added another panal class to hold the info for the second bar chart and
altered the code to use bar chart instead of histogram.
